Product displaying is now moved to Product Controller

+++ home.php
@@ -12,11 +12,15 @@ include_once './classes/product.controller.php';
 
                 <nav>
                     <ul>
-                        <?php
-                            if(isset($_SESSION["userid"])){
-                        ?>
+                        <?php if(isset($_SESSION["userid"])){ ?>
+                        <?php if($_SESSION["isAdmin"] == 1){ ?>
                             <li> <a href="/product/add"> Add </a> </li>
                             <li> <input type="submit" value="Mass Delete" name="deleteButton" onclick="<?php ProductController::delete();?>"> </li>
+                            <!-- Order status - possibly -->
+                        <?php }else{ ?>
+                            <li> <a id="profileButton" href="/profile"> <?php echo $_SESSION["name"] ?> </a> </li>
+                            <li> <a id="cartButton" href="/cart"> Cart </a> </li>
+                        <?php } ?>
                             <li> <input type="submit" value="Log Out" name="logoutbutton" onclick="<?php User::logOut();?>"> </li>
                         <?php
                             }else{
@@ -33,71 +37,9 @@ include_once './classes/product.controller.php';
     
         <div class= container>
             <?php
-                $result = Disk::selectFromDB();
-                if($result!=null){
-                    while($row = $result->fetch_assoc()){
-                        $sku = $row["SKU"];
-                        $name = $row["Name"];
-                        $price = $row["Price"];
-                        $size = $row["Size"];
-            ?>
-                    <div class="box">
-                        <input type="checkbox" name="deleteD[]" value="<?= $sku ?>">
-                        <p><?= $sku ?> </p>
-                        <p><?= $name ?> </p>
-                        <p><?= $price." $" ?> </p>
-                        <p><?= "Size: ".$size." MB "?></p>
-                    </div>
-            <?php
-                }
-            }
-            ?>
-            
-
-
-            <?php
-                $result = Book::selectFromDB();
-                if($result!=null){
-                    while($row = $result->fetch_assoc()){
-                        $sku = $row["SKU"];
-                        $name = $row["Name"];
-                        $price = $row["Price"];
-                        $weight = $row["Weight"];
-            ?>
-                    <div class="box">
-                        <input type="checkbox" name="deleteB[]" value="<?= $sku ?>">
-                        <p><?= $sku ?> </p>
-                        <p><?= $name ?> </p>
-                        <p><?= $price." $" ?> </p>
-                        <p><?= "Weight: ".$weight." KG "?></p>
-                    </div>
-            <?php
-                }
-            }
-            ?>
-
-
-            <?php
-                $result = Furniture::selectFromDB();
-                if($result!=null){
-                    while($row = $result->fetch_assoc()){
-                        $sku = $row["SKU"];
-                        $name = $row["Name"];
-                        $price = $row["Price"];
-                        $width = $row["Width"];
-                        $height = $row["Height"];
-                        $length = $row["Length"];
-            ?>
-                    <div class="box">
-                        <input type="checkbox" name="deleteF[]" value="<?= $sku ?>">
-                        <p><?= $sku ?> </p>
-                        <p><?= $name ?> </p>
-                        <p><?= $price." $" ?> </p>
-                        <p><?= "Dimention: ".$height."x".$width."x".$length?></p>
-                    </div>
-            <?php
-                }
-            }
+                Disk::show();
+                Book::show();
+                Furniture::show();
             ?>
